Commit now split file
