+++ src/ProgressOnderwijsUtils/Html/HtmlSpec.AttributeNameInterfaces.Generated.cs
@@ -1,0 +1,127 @@
+#nullable enable
+using static ProgressOnderwijsUtils.Html.AttributeNameInterfaces;
+
+namespace ProgressOnderwijsUtils.Html;
+
+public static class AttributeNameInterfaces
+{
+    public interface IHasAttr_href { }
+    public interface IHasAttr_target { }
+    public interface IHasAttr_download { }
+    public interface IHasAttr_ping { }
+    public interface IHasAttr_rel { }
+    public interface IHasAttr_hreflang { }
+    public interface IHasAttr_type { }
+    public interface IHasAttr_referrerpolicy { }
+    public interface IHasAttr_alt { }
+    public interface IHasAttr_coords { }
+    public interface IHasAttr_shape { }
+    public interface IHasAttr_src { }
+    public interface IHasAttr_crossorigin { }
+    public interface IHasAttr_preload { }
+    public interface IHasAttr_autoplay { }
+    public interface IHasAttr_loop { }
+    public interface IHasAttr_muted { }
+    public interface IHasAttr_controls { }
+    public interface IHasAttr_cite { }
+    public interface IHasAttr_onafterprint { }
+    public interface IHasAttr_onbeforeprint { }
+    public interface IHasAttr_onbeforeunload { }
+    public interface IHasAttr_onhashchange { }
+    public interface IHasAttr_onlanguagechange { }
+    public interface IHasAttr_onmessage { }
+    public interface IHasAttr_onmessageerror { }
+    public interface IHasAttr_onoffline { }
+    public interface IHasAttr_ononline { }
+    public interface IHasAttr_onpagehide { }
+    public interface IHasAttr_onpageshow { }
+    public interface IHasAttr_onpopstate { }
+    public interface IHasAttr_onrejectionhandled { }
+    public interface IHasAttr_onstorage { }
+    public interface IHasAttr_onunhandledrejection { }
+    public interface IHasAttr_onunload { }
+    public interface IHasAttr_disabled { }
+    public interface IHasAttr_form { }
+    public interface IHasAttr_formaction { }
+    public interface IHasAttr_formenctype { }
+    public interface IHasAttr_formmethod { }
+    public interface IHasAttr_formnovalidate { }
+    public interface IHasAttr_formtarget { }
+    public interface IHasAttr_name { }
+    public interface IHasAttr_popovertarget { }
+    public interface IHasAttr_popovertargetaction { }
+    public interface IHasAttr_value { }
+    public interface IHasAttr_width { }
+    public interface IHasAttr_height { }
+    public interface IHasAttr_span { }
+    public interface IHasAttr_datetime { }
+    public interface IHasAttr_open { }
+    public interface IHasAttr_accept_charset { }
+    public interface IHasAttr_action { }
+    public interface IHasAttr_autocomplete { }
+    public interface IHasAttr_enctype { }
+    public interface IHasAttr_method { }
+    public interface IHasAttr_novalidate { }
+    public interface IHasAttr_manifest { }
+    public interface IHasAttr_srcdoc { }
+    public interface IHasAttr_sandbox { }
+    public interface IHasAttr_allow { }
+    public interface IHasAttr_allowfullscreen { }
+    public interface IHasAttr_loading { }
+    public interface IHasAttr_srcset { }
+    public interface IHasAttr_sizes { }
+    public interface IHasAttr_usemap { }
+    public interface IHasAttr_ismap { }
+    public interface IHasAttr_decoding { }
+    public interface IHasAttr_fetchpriority { }
+    public interface IHasAttr_accept { }
+    public interface IHasAttr_checked { }
+    public interface IHasAttr_dirname { }
+    public interface IHasAttr_list { }
+    public interface IHasAttr_max { }
+    public interface IHasAttr_maxlength { }
+    public interface IHasAttr_min { }
+    public interface IHasAttr_minlength { }
+    public interface IHasAttr_multiple { }
+    public interface IHasAttr_pattern { }
+    public interface IHasAttr_placeholder { }
+    public interface IHasAttr_readonly { }
+    public interface IHasAttr_required { }
+    public interface IHasAttr_size { }
+    public interface IHasAttr_step { }
+    public interface IHasAttr_for { }
+    public interface IHasAttr_as { }
+    public interface IHasAttr_media { }
+    public interface IHasAttr_imagesrcset { }
+    public interface IHasAttr_imagesizes { }
+    public interface IHasAttr_integrity { }
+    public interface IHasAttr_blocking { }
+    public interface IHasAttr_color { }
+    public interface IHasAttr_http_equiv { }
+    public interface IHasAttr_content { }
+    public interface IHasAttr_charset { }
+    public interface IHasAttr_low { }
+    public interface IHasAttr_high { }
+    public interface IHasAttr_optimum { }
+    public interface IHasAttr_data { }
+    public interface IHasAttr_reversed { }
+    public interface IHasAttr_start { }
+    public interface IHasAttr_label { }
+    public interface IHasAttr_selected { }
+    public interface IHasAttr_nomodule { }
+    public interface IHasAttr_async { }
+    public interface IHasAttr_defer { }
+    public interface IHasAttr_colspan { }
+    public interface IHasAttr_rowspan { }
+    public interface IHasAttr_headers { }
+    public interface IHasAttr_cols { }
+    public interface IHasAttr_rows { }
+    public interface IHasAttr_wrap { }
+    public interface IHasAttr_scope { }
+    public interface IHasAttr_abbr { }
+    public interface IHasAttr_default { }
+    public interface IHasAttr_kind { }
+    public interface IHasAttr_srclang { }
+    public interface IHasAttr_poster { }
+    public interface IHasAttr_playsinline { }
+}
