+++ src/ProgressOnderwijsUtils/Html/HtmlSpec.HtmlTagKinds.Generated.cs
@@ -1871,909 +1871,3 @@ public static class HtmlTagKinds
         public static HtmlFragment operator +(string head, WBR tail) => HtmlFragment.Fragment(head, HtmlFragment.Element(tail));
     }
 }
-
-public static class Tags
-{
-
-    ///<summary>Hyperlink. See: <a href="https://html.spec.whatwg.org/#the-a-element">https://html.spec.whatwg.org/#the-a-element</a><br /></summary>
-    public static readonly HtmlTagKinds.A _a = new HtmlTagKinds.A();
-
-    ///<summary>Abbreviation. See: <a href="https://html.spec.whatwg.org/#the-abbr-element">https://html.spec.whatwg.org/#the-abbr-element</a><br /></summary>
-    public static readonly HtmlTagKinds.ABBR _abbr = new HtmlTagKinds.ABBR();
-
-    ///<summary>Contact information for a page or article element. See: <a href="https://html.spec.whatwg.org/#the-address-element">https://html.spec.whatwg.org/#the-address-element</a><br /></summary>
-    public static readonly HtmlTagKinds.ADDRESS _address = new HtmlTagKinds.ADDRESS();
-
-    ///<summary>Hyperlink or dead area on an image map. See: <a href="https://html.spec.whatwg.org/#the-area-element">https://html.spec.whatwg.org/#the-area-element</a><br /></summary>
-    public static readonly HtmlTagKinds.AREA _area = new HtmlTagKinds.AREA();
-
-    ///<summary>Self-contained syndicatable or reusable composition. See: <a href="https://html.spec.whatwg.org/#the-article-element">https://html.spec.whatwg.org/#the-article-element</a><br /></summary>
-    public static readonly HtmlTagKinds.ARTICLE _article = new HtmlTagKinds.ARTICLE();
-
-    ///<summary>Sidebar for tangentially related content. See: <a href="https://html.spec.whatwg.org/#the-aside-element">https://html.spec.whatwg.org/#the-aside-element</a><br /></summary>
-    public static readonly HtmlTagKinds.ASIDE _aside = new HtmlTagKinds.ASIDE();
-
-    ///<summary>Audio player. See: <a href="https://html.spec.whatwg.org/#the-audio-element">https://html.spec.whatwg.org/#the-audio-element</a><br /></summary>
-    public static readonly HtmlTagKinds.AUDIO _audio = new HtmlTagKinds.AUDIO();
-
-    ///<summary>Keywords. See: <a href="https://html.spec.whatwg.org/#the-b-element">https://html.spec.whatwg.org/#the-b-element</a><br /></summary>
-    public static readonly HtmlTagKinds.B _b = new HtmlTagKinds.B();
-
-    ///<summary>Base URL and default target navigable for hyperlinks and forms. See: <a href="https://html.spec.whatwg.org/#the-base-element">https://html.spec.whatwg.org/#the-base-element</a><br /></summary>
-    public static readonly HtmlTagKinds.BASE _base = new HtmlTagKinds.BASE();
-
-    ///<summary>Text directionality isolation. See: <a href="https://html.spec.whatwg.org/#the-bdi-element">https://html.spec.whatwg.org/#the-bdi-element</a><br /></summary>
-    public static readonly HtmlTagKinds.BDI _bdi = new HtmlTagKinds.BDI();
-
-    ///<summary>Text directionality formatting. See: <a href="https://html.spec.whatwg.org/#the-bdo-element">https://html.spec.whatwg.org/#the-bdo-element</a><br /></summary>
-    public static readonly HtmlTagKinds.BDO _bdo = new HtmlTagKinds.BDO();
-
-    ///<summary>A section quoted from another source. See: <a href="https://html.spec.whatwg.org/#the-blockquote-element">https://html.spec.whatwg.org/#the-blockquote-element</a><br /></summary>
-    public static readonly HtmlTagKinds.BLOCKQUOTE _blockquote = new HtmlTagKinds.BLOCKQUOTE();
-
-    ///<summary>Document body. See: <a href="https://html.spec.whatwg.org/#the-body-element">https://html.spec.whatwg.org/#the-body-element</a><br /></summary>
-    public static readonly HtmlTagKinds.BODY _body = new HtmlTagKinds.BODY();
-
-    ///<summary>Line break, e.g. in poem or postal address. See: <a href="https://html.spec.whatwg.org/#the-br-element">https://html.spec.whatwg.org/#the-br-element</a><br /></summary>
-    public static readonly HtmlTagKinds.BR _br = new HtmlTagKinds.BR();
-
-    ///<summary>Button control. See: <a href="https://html.spec.whatwg.org/#the-button-element">https://html.spec.whatwg.org/#the-button-element</a><br /></summary>
-    public static readonly HtmlTagKinds.BUTTON _button = new HtmlTagKinds.BUTTON();
-
-    ///<summary>Scriptable bitmap canvas. See: <a href="https://html.spec.whatwg.org/#the-canvas-element">https://html.spec.whatwg.org/#the-canvas-element</a><br /></summary>
-    public static readonly HtmlTagKinds.CANVAS _canvas = new HtmlTagKinds.CANVAS();
-
-    ///<summary>Table caption. See: <a href="https://html.spec.whatwg.org/#the-caption-element">https://html.spec.whatwg.org/#the-caption-element</a><br /></summary>
-    public static readonly HtmlTagKinds.CAPTION _caption = new HtmlTagKinds.CAPTION();
-
-    ///<summary>Title of a work. See: <a href="https://html.spec.whatwg.org/#the-cite-element">https://html.spec.whatwg.org/#the-cite-element</a><br /></summary>
-    public static readonly HtmlTagKinds.CITE _cite = new HtmlTagKinds.CITE();
-
-    ///<summary>Computer code. See: <a href="https://html.spec.whatwg.org/#the-code-element">https://html.spec.whatwg.org/#the-code-element</a><br /></summary>
-    public static readonly HtmlTagKinds.CODE _code = new HtmlTagKinds.CODE();
-
-    ///<summary>Table column. See: <a href="https://html.spec.whatwg.org/#the-col-element">https://html.spec.whatwg.org/#the-col-element</a><br /></summary>
-    public static readonly HtmlTagKinds.COL _col = new HtmlTagKinds.COL();
-
-    ///<summary>Group of columns in a table. See: <a href="https://html.spec.whatwg.org/#the-colgroup-element">https://html.spec.whatwg.org/#the-colgroup-element</a><br /></summary>
-    public static readonly HtmlTagKinds.COLGROUP _colgroup = new HtmlTagKinds.COLGROUP();
-
-    ///<summary>Machine-readable equivalent. See: <a href="https://html.spec.whatwg.org/#the-data-element">https://html.spec.whatwg.org/#the-data-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DATA _data = new HtmlTagKinds.DATA();
-
-    ///<summary>Container for options for combo box control. See: <a href="https://html.spec.whatwg.org/#the-datalist-element">https://html.spec.whatwg.org/#the-datalist-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DATALIST _datalist = new HtmlTagKinds.DATALIST();
-
-    ///<summary>Content for corresponding dt element(s). See: <a href="https://html.spec.whatwg.org/#the-dd-element">https://html.spec.whatwg.org/#the-dd-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DD _dd = new HtmlTagKinds.DD();
-
-    ///<summary>A removal from the document. See: <a href="https://html.spec.whatwg.org/#the-del-element">https://html.spec.whatwg.org/#the-del-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DEL _del = new HtmlTagKinds.DEL();
-
-    ///<summary>Disclosure control for hiding details. See: <a href="https://html.spec.whatwg.org/#the-details-element">https://html.spec.whatwg.org/#the-details-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DETAILS _details = new HtmlTagKinds.DETAILS();
-
-    ///<summary>Defining instance. See: <a href="https://html.spec.whatwg.org/#the-dfn-element">https://html.spec.whatwg.org/#the-dfn-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DFN _dfn = new HtmlTagKinds.DFN();
-
-    ///<summary>Dialog box or window. See: <a href="https://html.spec.whatwg.org/#the-dialog-element">https://html.spec.whatwg.org/#the-dialog-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DIALOG _dialog = new HtmlTagKinds.DIALOG();
-
-    ///<summary>Generic flow container, or container for name-value groups in dl elements. See: <a href="https://html.spec.whatwg.org/#the-div-element">https://html.spec.whatwg.org/#the-div-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DIV _div = new HtmlTagKinds.DIV();
-
-    ///<summary>Association list consisting of zero or more name-value groups. See: <a href="https://html.spec.whatwg.org/#the-dl-element">https://html.spec.whatwg.org/#the-dl-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DL _dl = new HtmlTagKinds.DL();
-
-    ///<summary>Legend for corresponding dd element(s). See: <a href="https://html.spec.whatwg.org/#the-dt-element">https://html.spec.whatwg.org/#the-dt-element</a><br /></summary>
-    public static readonly HtmlTagKinds.DT _dt = new HtmlTagKinds.DT();
-
-    ///<summary>Stress emphasis. See: <a href="https://html.spec.whatwg.org/#the-em-element">https://html.spec.whatwg.org/#the-em-element</a><br /></summary>
-    public static readonly HtmlTagKinds.EM _em = new HtmlTagKinds.EM();
-
-    ///<summary>Plugin. See: <a href="https://html.spec.whatwg.org/#the-embed-element">https://html.spec.whatwg.org/#the-embed-element</a><br /></summary>
-    public static readonly HtmlTagKinds.EMBED _embed = new HtmlTagKinds.EMBED();
-
-    ///<summary>Group of form controls. See: <a href="https://html.spec.whatwg.org/#the-fieldset-element">https://html.spec.whatwg.org/#the-fieldset-element</a><br /></summary>
-    public static readonly HtmlTagKinds.FIELDSET _fieldset = new HtmlTagKinds.FIELDSET();
-
-    ///<summary>Caption for figure. See: <a href="https://html.spec.whatwg.org/#the-figcaption-element">https://html.spec.whatwg.org/#the-figcaption-element</a><br /></summary>
-    public static readonly HtmlTagKinds.FIGCAPTION _figcaption = new HtmlTagKinds.FIGCAPTION();
-
-    ///<summary>Figure with optional caption. See: <a href="https://html.spec.whatwg.org/#the-figure-element">https://html.spec.whatwg.org/#the-figure-element</a><br /></summary>
-    public static readonly HtmlTagKinds.FIGURE _figure = new HtmlTagKinds.FIGURE();
-
-    ///<summary>Footer for a page or section. See: <a href="https://html.spec.whatwg.org/#the-footer-element">https://html.spec.whatwg.org/#the-footer-element</a><br /></summary>
-    public static readonly HtmlTagKinds.FOOTER _footer = new HtmlTagKinds.FOOTER();
-
-    ///<summary>User-submittable form. See: <a href="https://html.spec.whatwg.org/#the-form-element">https://html.spec.whatwg.org/#the-form-element</a><br /></summary>
-    public static readonly HtmlTagKinds.FORM _form = new HtmlTagKinds.FORM();
-
-    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.H1 _h1 = new HtmlTagKinds.H1();
-
-    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.H2 _h2 = new HtmlTagKinds.H2();
-
-    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.H3 _h3 = new HtmlTagKinds.H3();
-
-    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.H4 _h4 = new HtmlTagKinds.H4();
-
-    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.H5 _h5 = new HtmlTagKinds.H5();
-
-    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.H6 _h6 = new HtmlTagKinds.H6();
-
-    ///<summary>Container for document metadata. See: <a href="https://html.spec.whatwg.org/#the-head-element">https://html.spec.whatwg.org/#the-head-element</a><br /></summary>
-    public static readonly HtmlTagKinds.HEAD _head = new HtmlTagKinds.HEAD();
-
-    ///<summary>Introductory or navigational aids for a page or section. See: <a href="https://html.spec.whatwg.org/#the-header-element">https://html.spec.whatwg.org/#the-header-element</a><br /></summary>
-    public static readonly HtmlTagKinds.HEADER _header = new HtmlTagKinds.HEADER();
-
-    ///<summary>Heading container. See: <a href="https://html.spec.whatwg.org/#the-hgroup-element">https://html.spec.whatwg.org/#the-hgroup-element</a><br /></summary>
-    public static readonly HtmlTagKinds.HGROUP _hgroup = new HtmlTagKinds.HGROUP();
-
-    ///<summary>Thematic break. See: <a href="https://html.spec.whatwg.org/#the-hr-element">https://html.spec.whatwg.org/#the-hr-element</a><br /></summary>
-    public static readonly HtmlTagKinds.HR _hr = new HtmlTagKinds.HR();
-
-    ///<summary>Root element. See: <a href="https://html.spec.whatwg.org/#the-html-element">https://html.spec.whatwg.org/#the-html-element</a><br /></summary>
-    public static readonly HtmlTagKinds.HTML _html = new HtmlTagKinds.HTML();
-
-    ///<summary>Alternate voice. See: <a href="https://html.spec.whatwg.org/#the-i-element">https://html.spec.whatwg.org/#the-i-element</a><br /></summary>
-    public static readonly HtmlTagKinds.I _i = new HtmlTagKinds.I();
-
-    ///<summary>Child navigable. See: <a href="https://html.spec.whatwg.org/#the-iframe-element">https://html.spec.whatwg.org/#the-iframe-element</a><br /></summary>
-    public static readonly HtmlTagKinds.IFRAME _iframe = new HtmlTagKinds.IFRAME();
-
-    ///<summary>Image. See: <a href="https://html.spec.whatwg.org/#the-img-element">https://html.spec.whatwg.org/#the-img-element</a><br /></summary>
-    public static readonly HtmlTagKinds.IMG _img = new HtmlTagKinds.IMG();
-
-    ///<summary>Form control. See: <a href="https://html.spec.whatwg.org/#the-input-element">https://html.spec.whatwg.org/#the-input-element</a><br /></summary>
-    public static readonly HtmlTagKinds.INPUT _input = new HtmlTagKinds.INPUT();
-
-    ///<summary>An addition to the document. See: <a href="https://html.spec.whatwg.org/#the-ins-element">https://html.spec.whatwg.org/#the-ins-element</a><br /></summary>
-    public static readonly HtmlTagKinds.INS _ins = new HtmlTagKinds.INS();
-
-    ///<summary>User input. See: <a href="https://html.spec.whatwg.org/#the-kbd-element">https://html.spec.whatwg.org/#the-kbd-element</a><br /></summary>
-    public static readonly HtmlTagKinds.KBD _kbd = new HtmlTagKinds.KBD();
-
-    ///<summary>Caption for a form control. See: <a href="https://html.spec.whatwg.org/#the-label-element">https://html.spec.whatwg.org/#the-label-element</a><br /></summary>
-    public static readonly HtmlTagKinds.LABEL _label = new HtmlTagKinds.LABEL();
-
-    ///<summary>Caption for fieldset. See: <a href="https://html.spec.whatwg.org/#the-legend-element">https://html.spec.whatwg.org/#the-legend-element</a><br /></summary>
-    public static readonly HtmlTagKinds.LEGEND _legend = new HtmlTagKinds.LEGEND();
-
-    ///<summary>List item. See: <a href="https://html.spec.whatwg.org/#the-li-element">https://html.spec.whatwg.org/#the-li-element</a><br /></summary>
-    public static readonly HtmlTagKinds.LI _li = new HtmlTagKinds.LI();
-
-    ///<summary>Link metadata. See: <a href="https://html.spec.whatwg.org/#the-link-element">https://html.spec.whatwg.org/#the-link-element</a><br /></summary>
-    public static readonly HtmlTagKinds.LINK _link = new HtmlTagKinds.LINK();
-
-    ///<summary>Container for the dominant contents of the document. See: <a href="https://html.spec.whatwg.org/#the-main-element">https://html.spec.whatwg.org/#the-main-element</a><br /></summary>
-    public static readonly HtmlTagKinds.MAIN _main = new HtmlTagKinds.MAIN();
-
-    ///<summary>Image map. See: <a href="https://html.spec.whatwg.org/#the-map-element">https://html.spec.whatwg.org/#the-map-element</a><br /></summary>
-    public static readonly HtmlTagKinds.MAP _map = new HtmlTagKinds.MAP();
-
-    ///<summary>Highlight. See: <a href="https://html.spec.whatwg.org/#the-mark-element">https://html.spec.whatwg.org/#the-mark-element</a><br /></summary>
-    public static readonly HtmlTagKinds.MARK _mark = new HtmlTagKinds.MARK();
-
-    ///<summary>Menu of commands. See: <a href="https://html.spec.whatwg.org/#the-menu-element">https://html.spec.whatwg.org/#the-menu-element</a><br /></summary>
-    public static readonly HtmlTagKinds.MENU _menu = new HtmlTagKinds.MENU();
-
-    ///<summary>Text metadata. See: <a href="https://html.spec.whatwg.org/#the-meta-element">https://html.spec.whatwg.org/#the-meta-element</a><br /></summary>
-    public static readonly HtmlTagKinds.META _meta = new HtmlTagKinds.META();
-
-    ///<summary>Gauge. See: <a href="https://html.spec.whatwg.org/#the-meter-element">https://html.spec.whatwg.org/#the-meter-element</a><br /></summary>
-    public static readonly HtmlTagKinds.METER _meter = new HtmlTagKinds.METER();
-
-    ///<summary>Section with navigational links. See: <a href="https://html.spec.whatwg.org/#the-nav-element">https://html.spec.whatwg.org/#the-nav-element</a><br /></summary>
-    public static readonly HtmlTagKinds.NAV _nav = new HtmlTagKinds.NAV();
-
-    ///<summary>Fallback content for script. See: <a href="https://html.spec.whatwg.org/#the-noscript-element">https://html.spec.whatwg.org/#the-noscript-element</a><br /></summary>
-    public static readonly HtmlTagKinds.NOSCRIPT _noscript = new HtmlTagKinds.NOSCRIPT();
-
-    ///<summary>Image, child navigable, or plugin. See: <a href="https://html.spec.whatwg.org/#the-object-element">https://html.spec.whatwg.org/#the-object-element</a><br /></summary>
-    public static readonly HtmlTagKinds.OBJECT _object = new HtmlTagKinds.OBJECT();
-
-    ///<summary>Ordered list. See: <a href="https://html.spec.whatwg.org/#the-ol-element">https://html.spec.whatwg.org/#the-ol-element</a><br /></summary>
-    public static readonly HtmlTagKinds.OL _ol = new HtmlTagKinds.OL();
-
-    ///<summary>Group of options in a list box. See: <a href="https://html.spec.whatwg.org/#the-optgroup-element">https://html.spec.whatwg.org/#the-optgroup-element</a><br /></summary>
-    public static readonly HtmlTagKinds.OPTGROUP _optgroup = new HtmlTagKinds.OPTGROUP();
-
-    ///<summary>Option in a list box or combo box control. See: <a href="https://html.spec.whatwg.org/#the-option-element">https://html.spec.whatwg.org/#the-option-element</a><br /></summary>
-    public static readonly HtmlTagKinds.OPTION _option = new HtmlTagKinds.OPTION();
-
-    ///<summary>Calculated output value. See: <a href="https://html.spec.whatwg.org/#the-output-element">https://html.spec.whatwg.org/#the-output-element</a><br /></summary>
-    public static readonly HtmlTagKinds.OUTPUT _output = new HtmlTagKinds.OUTPUT();
-
-    ///<summary>Paragraph. See: <a href="https://html.spec.whatwg.org/#the-p-element">https://html.spec.whatwg.org/#the-p-element</a><br /></summary>
-    public static readonly HtmlTagKinds.P _p = new HtmlTagKinds.P();
-
-    ///<summary>Image. See: <a href="https://html.spec.whatwg.org/#the-picture-element">https://html.spec.whatwg.org/#the-picture-element</a><br /></summary>
-    public static readonly HtmlTagKinds.PICTURE _picture = new HtmlTagKinds.PICTURE();
-
-    ///<summary>Block of preformatted text. See: <a href="https://html.spec.whatwg.org/#the-pre-element">https://html.spec.whatwg.org/#the-pre-element</a><br /></summary>
-    public static readonly HtmlTagKinds.PRE _pre = new HtmlTagKinds.PRE();
-
-    ///<summary>Progress bar. See: <a href="https://html.spec.whatwg.org/#the-progress-element">https://html.spec.whatwg.org/#the-progress-element</a><br /></summary>
-    public static readonly HtmlTagKinds.PROGRESS _progress = new HtmlTagKinds.PROGRESS();
-
-    ///<summary>Quotation. See: <a href="https://html.spec.whatwg.org/#the-q-element">https://html.spec.whatwg.org/#the-q-element</a><br /></summary>
-    public static readonly HtmlTagKinds.Q _q = new HtmlTagKinds.Q();
-
-    ///<summary>Parenthesis for ruby annotation text. See: <a href="https://html.spec.whatwg.org/#the-rp-element">https://html.spec.whatwg.org/#the-rp-element</a><br /></summary>
-    public static readonly HtmlTagKinds.RP _rp = new HtmlTagKinds.RP();
-
-    ///<summary>Ruby annotation text. See: <a href="https://html.spec.whatwg.org/#the-rt-element">https://html.spec.whatwg.org/#the-rt-element</a><br /></summary>
-    public static readonly HtmlTagKinds.RT _rt = new HtmlTagKinds.RT();
-
-    ///<summary>Ruby annotation(s). See: <a href="https://html.spec.whatwg.org/#the-ruby-element">https://html.spec.whatwg.org/#the-ruby-element</a><br /></summary>
-    public static readonly HtmlTagKinds.RUBY _ruby = new HtmlTagKinds.RUBY();
-
-    ///<summary>Inaccurate text. See: <a href="https://html.spec.whatwg.org/#the-s-element">https://html.spec.whatwg.org/#the-s-element</a><br /></summary>
-    public static readonly HtmlTagKinds.S _s = new HtmlTagKinds.S();
-
-    ///<summary>Computer output. See: <a href="https://html.spec.whatwg.org/#the-samp-element">https://html.spec.whatwg.org/#the-samp-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SAMP _samp = new HtmlTagKinds.SAMP();
-
-    ///<summary>Embedded script. See: <a href="https://html.spec.whatwg.org/#the-script-element">https://html.spec.whatwg.org/#the-script-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SCRIPT _script = new HtmlTagKinds.SCRIPT();
-
-    ///<summary>Container for search controls. See: <a href="https://html.spec.whatwg.org/#the-search-element">https://html.spec.whatwg.org/#the-search-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SEARCH _search = new HtmlTagKinds.SEARCH();
-
-    ///<summary>Generic document or application section. See: <a href="https://html.spec.whatwg.org/#the-section-element">https://html.spec.whatwg.org/#the-section-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SECTION _section = new HtmlTagKinds.SECTION();
-
-    ///<summary>List box control. See: <a href="https://html.spec.whatwg.org/#the-select-element">https://html.spec.whatwg.org/#the-select-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SELECT _select = new HtmlTagKinds.SELECT();
-
-    ///<summary>Shadow tree slot. See: <a href="https://html.spec.whatwg.org/#the-slot-element">https://html.spec.whatwg.org/#the-slot-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SLOT _slot = new HtmlTagKinds.SLOT();
-
-    ///<summary>Side comment. See: <a href="https://html.spec.whatwg.org/#the-small-element">https://html.spec.whatwg.org/#the-small-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SMALL _small = new HtmlTagKinds.SMALL();
-
-    ///<summary>Image source for img or media source for video or audio. See: <a href="https://html.spec.whatwg.org/#the-source-element">https://html.spec.whatwg.org/#the-source-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SOURCE _source = new HtmlTagKinds.SOURCE();
-
-    ///<summary>Generic phrasing container. See: <a href="https://html.spec.whatwg.org/#the-span-element">https://html.spec.whatwg.org/#the-span-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SPAN _span = new HtmlTagKinds.SPAN();
-
-    ///<summary>Importance. See: <a href="https://html.spec.whatwg.org/#the-strong-element">https://html.spec.whatwg.org/#the-strong-element</a><br /></summary>
-    public static readonly HtmlTagKinds.STRONG _strong = new HtmlTagKinds.STRONG();
-
-    ///<summary>Embedded styling information. See: <a href="https://html.spec.whatwg.org/#the-style-element">https://html.spec.whatwg.org/#the-style-element</a><br /></summary>
-    public static readonly HtmlTagKinds.STYLE _style = new HtmlTagKinds.STYLE();
-
-    ///<summary>Subscript. See: <a href="https://html.spec.whatwg.org/#the-sub-and-sup-elements">https://html.spec.whatwg.org/#the-sub-and-sup-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.SUB _sub = new HtmlTagKinds.SUB();
-
-    ///<summary>Caption for details. See: <a href="https://html.spec.whatwg.org/#the-summary-element">https://html.spec.whatwg.org/#the-summary-element</a><br /></summary>
-    public static readonly HtmlTagKinds.SUMMARY _summary = new HtmlTagKinds.SUMMARY();
-
-    ///<summary>Superscript. See: <a href="https://html.spec.whatwg.org/#the-sub-and-sup-elements">https://html.spec.whatwg.org/#the-sub-and-sup-elements</a><br /></summary>
-    public static readonly HtmlTagKinds.SUP _sup = new HtmlTagKinds.SUP();
-
-    ///<summary>Table. See: <a href="https://html.spec.whatwg.org/#the-table-element">https://html.spec.whatwg.org/#the-table-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TABLE _table = new HtmlTagKinds.TABLE();
-
-    ///<summary>Group of rows in a table. See: <a href="https://html.spec.whatwg.org/#the-tbody-element">https://html.spec.whatwg.org/#the-tbody-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TBODY _tbody = new HtmlTagKinds.TBODY();
-
-    ///<summary>Table cell. See: <a href="https://html.spec.whatwg.org/#the-td-element">https://html.spec.whatwg.org/#the-td-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TD _td = new HtmlTagKinds.TD();
-
-    ///<summary>Template. See: <a href="https://html.spec.whatwg.org/#the-template-element">https://html.spec.whatwg.org/#the-template-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TEMPLATE _template = new HtmlTagKinds.TEMPLATE();
-
-    ///<summary>Multiline text controls. See: <a href="https://html.spec.whatwg.org/#the-textarea-element">https://html.spec.whatwg.org/#the-textarea-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TEXTAREA _textarea = new HtmlTagKinds.TEXTAREA();
-
-    ///<summary>Group of footer rows in a table. See: <a href="https://html.spec.whatwg.org/#the-tfoot-element">https://html.spec.whatwg.org/#the-tfoot-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TFOOT _tfoot = new HtmlTagKinds.TFOOT();
-
-    ///<summary>Table header cell. See: <a href="https://html.spec.whatwg.org/#the-th-element">https://html.spec.whatwg.org/#the-th-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TH _th = new HtmlTagKinds.TH();
-
-    ///<summary>Group of heading rows in a table. See: <a href="https://html.spec.whatwg.org/#the-thead-element">https://html.spec.whatwg.org/#the-thead-element</a><br /></summary>
-    public static readonly HtmlTagKinds.THEAD _thead = new HtmlTagKinds.THEAD();
-
-    ///<summary>Machine-readable equivalent of date- or time-related data. See: <a href="https://html.spec.whatwg.org/#the-time-element">https://html.spec.whatwg.org/#the-time-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TIME _time = new HtmlTagKinds.TIME();
-
-    ///<summary>Document title. See: <a href="https://html.spec.whatwg.org/#the-title-element">https://html.spec.whatwg.org/#the-title-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TITLE _title = new HtmlTagKinds.TITLE();
-
-    ///<summary>Table row. See: <a href="https://html.spec.whatwg.org/#the-tr-element">https://html.spec.whatwg.org/#the-tr-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TR _tr = new HtmlTagKinds.TR();
-
-    ///<summary>Timed text track. See: <a href="https://html.spec.whatwg.org/#the-track-element">https://html.spec.whatwg.org/#the-track-element</a><br /></summary>
-    public static readonly HtmlTagKinds.TRACK _track = new HtmlTagKinds.TRACK();
-
-    ///<summary>Unarticulated annotation. See: <a href="https://html.spec.whatwg.org/#the-u-element">https://html.spec.whatwg.org/#the-u-element</a><br /></summary>
-    public static readonly HtmlTagKinds.U _u = new HtmlTagKinds.U();
-
-    ///<summary>List. See: <a href="https://html.spec.whatwg.org/#the-ul-element">https://html.spec.whatwg.org/#the-ul-element</a><br /></summary>
-    public static readonly HtmlTagKinds.UL _ul = new HtmlTagKinds.UL();
-
-    ///<summary>Variable. See: <a href="https://html.spec.whatwg.org/#the-var-element">https://html.spec.whatwg.org/#the-var-element</a><br /></summary>
-    public static readonly HtmlTagKinds.VAR _var = new HtmlTagKinds.VAR();
-
-    ///<summary>Video player. See: <a href="https://html.spec.whatwg.org/#the-video-element">https://html.spec.whatwg.org/#the-video-element</a><br /></summary>
-    public static readonly HtmlTagKinds.VIDEO _video = new HtmlTagKinds.VIDEO();
-
-    ///<summary>Line breaking opportunity. See: <a href="https://html.spec.whatwg.org/#the-wbr-element">https://html.spec.whatwg.org/#the-wbr-element</a><br /></summary>
-    public static readonly HtmlTagKinds.WBR _wbr = new HtmlTagKinds.WBR();
-}
-
-public static class AttributeNameInterfaces
-{
-    public interface IHasAttr_href { }
-    public interface IHasAttr_target { }
-    public interface IHasAttr_download { }
-    public interface IHasAttr_ping { }
-    public interface IHasAttr_rel { }
-    public interface IHasAttr_hreflang { }
-    public interface IHasAttr_type { }
-    public interface IHasAttr_referrerpolicy { }
-    public interface IHasAttr_alt { }
-    public interface IHasAttr_coords { }
-    public interface IHasAttr_shape { }
-    public interface IHasAttr_src { }
-    public interface IHasAttr_crossorigin { }
-    public interface IHasAttr_preload { }
-    public interface IHasAttr_autoplay { }
-    public interface IHasAttr_loop { }
-    public interface IHasAttr_muted { }
-    public interface IHasAttr_controls { }
-    public interface IHasAttr_cite { }
-    public interface IHasAttr_onafterprint { }
-    public interface IHasAttr_onbeforeprint { }
-    public interface IHasAttr_onbeforeunload { }
-    public interface IHasAttr_onhashchange { }
-    public interface IHasAttr_onlanguagechange { }
-    public interface IHasAttr_onmessage { }
-    public interface IHasAttr_onmessageerror { }
-    public interface IHasAttr_onoffline { }
-    public interface IHasAttr_ononline { }
-    public interface IHasAttr_onpagehide { }
-    public interface IHasAttr_onpageshow { }
-    public interface IHasAttr_onpopstate { }
-    public interface IHasAttr_onrejectionhandled { }
-    public interface IHasAttr_onstorage { }
-    public interface IHasAttr_onunhandledrejection { }
-    public interface IHasAttr_onunload { }
-    public interface IHasAttr_disabled { }
-    public interface IHasAttr_form { }
-    public interface IHasAttr_formaction { }
-    public interface IHasAttr_formenctype { }
-    public interface IHasAttr_formmethod { }
-    public interface IHasAttr_formnovalidate { }
-    public interface IHasAttr_formtarget { }
-    public interface IHasAttr_name { }
-    public interface IHasAttr_popovertarget { }
-    public interface IHasAttr_popovertargetaction { }
-    public interface IHasAttr_value { }
-    public interface IHasAttr_width { }
-    public interface IHasAttr_height { }
-    public interface IHasAttr_span { }
-    public interface IHasAttr_datetime { }
-    public interface IHasAttr_open { }
-    public interface IHasAttr_accept_charset { }
-    public interface IHasAttr_action { }
-    public interface IHasAttr_autocomplete { }
-    public interface IHasAttr_enctype { }
-    public interface IHasAttr_method { }
-    public interface IHasAttr_novalidate { }
-    public interface IHasAttr_manifest { }
-    public interface IHasAttr_srcdoc { }
-    public interface IHasAttr_sandbox { }
-    public interface IHasAttr_allow { }
-    public interface IHasAttr_allowfullscreen { }
-    public interface IHasAttr_loading { }
-    public interface IHasAttr_srcset { }
-    public interface IHasAttr_sizes { }
-    public interface IHasAttr_usemap { }
-    public interface IHasAttr_ismap { }
-    public interface IHasAttr_decoding { }
-    public interface IHasAttr_fetchpriority { }
-    public interface IHasAttr_accept { }
-    public interface IHasAttr_checked { }
-    public interface IHasAttr_dirname { }
-    public interface IHasAttr_list { }
-    public interface IHasAttr_max { }
-    public interface IHasAttr_maxlength { }
-    public interface IHasAttr_min { }
-    public interface IHasAttr_minlength { }
-    public interface IHasAttr_multiple { }
-    public interface IHasAttr_pattern { }
-    public interface IHasAttr_placeholder { }
-    public interface IHasAttr_readonly { }
-    public interface IHasAttr_required { }
-    public interface IHasAttr_size { }
-    public interface IHasAttr_step { }
-    public interface IHasAttr_for { }
-    public interface IHasAttr_as { }
-    public interface IHasAttr_media { }
-    public interface IHasAttr_imagesrcset { }
-    public interface IHasAttr_imagesizes { }
-    public interface IHasAttr_integrity { }
-    public interface IHasAttr_blocking { }
-    public interface IHasAttr_color { }
-    public interface IHasAttr_http_equiv { }
-    public interface IHasAttr_content { }
-    public interface IHasAttr_charset { }
-    public interface IHasAttr_low { }
-    public interface IHasAttr_high { }
-    public interface IHasAttr_optimum { }
-    public interface IHasAttr_data { }
-    public interface IHasAttr_reversed { }
-    public interface IHasAttr_start { }
-    public interface IHasAttr_label { }
-    public interface IHasAttr_selected { }
-    public interface IHasAttr_nomodule { }
-    public interface IHasAttr_async { }
-    public interface IHasAttr_defer { }
-    public interface IHasAttr_colspan { }
-    public interface IHasAttr_rowspan { }
-    public interface IHasAttr_headers { }
-    public interface IHasAttr_cols { }
-    public interface IHasAttr_rows { }
-    public interface IHasAttr_wrap { }
-    public interface IHasAttr_scope { }
-    public interface IHasAttr_abbr { }
-    public interface IHasAttr_default { }
-    public interface IHasAttr_kind { }
-    public interface IHasAttr_srclang { }
-    public interface IHasAttr_poster { }
-    public interface IHasAttr_playsinline { }
-}
-
-public static class AttributeConstructionMethods
-{
-    public static THtmlTag _accesskey<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("accesskey", attrValue);
-    public static THtmlTag _autocapitalize<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("autocapitalize", attrValue);
-    public static THtmlTag _autofocus<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("autofocus", attrValue);
-    [Obsolete]
-    public static THtmlTag _class<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("class", attrValue);
-    public static THtmlTag _contenteditable<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("contenteditable", attrValue);
-    public static THtmlTag _dir<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("dir", attrValue);
-    public static THtmlTag _draggable<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("draggable", attrValue);
-    public static THtmlTag _enterkeyhint<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("enterkeyhint", attrValue);
-    public static THtmlTag _hidden<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("hidden", attrValue);
-    public static THtmlTag _id<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("id", attrValue);
-    public static THtmlTag _inert<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("inert", attrValue);
-    public static THtmlTag _inputmode<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("inputmode", attrValue);
-    public static THtmlTag _is<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("is", attrValue);
-    public static THtmlTag _itemid<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("itemid", attrValue);
-    public static THtmlTag _itemprop<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("itemprop", attrValue);
-    public static THtmlTag _itemref<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("itemref", attrValue);
-    public static THtmlTag _itemscope<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("itemscope", attrValue);
-    public static THtmlTag _itemtype<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("itemtype", attrValue);
-    public static THtmlTag _lang<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("lang", attrValue);
-    public static THtmlTag _nonce<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("nonce", attrValue);
-    public static THtmlTag _popover<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("popover", attrValue);
-    public static THtmlTag _slot<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("slot", attrValue);
-    public static THtmlTag _spellcheck<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("spellcheck", attrValue);
-    public static THtmlTag _style<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("style", attrValue);
-    public static THtmlTag _tabindex<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("tabindex", attrValue);
-    public static THtmlTag _title<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("title", attrValue);
-    public static THtmlTag _translate<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("translate", attrValue);
-    public static THtmlTag _href<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_href, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("href", attrValue);
-    public static THtmlTag _target<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_target, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("target", attrValue);
-    public static THtmlTag _download<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_download, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("download", attrValue);
-    public static THtmlTag _ping<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_ping, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("ping", attrValue);
-    public static THtmlTag _rel<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_rel, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("rel", attrValue);
-    public static THtmlTag _hreflang<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_hreflang, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("hreflang", attrValue);
-    public static THtmlTag _type<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_type, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("type", attrValue);
-    public static THtmlTag _referrerpolicy<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_referrerpolicy, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("referrerpolicy", attrValue);
-    public static THtmlTag _alt<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_alt, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("alt", attrValue);
-    public static THtmlTag _coords<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_coords, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("coords", attrValue);
-    public static THtmlTag _shape<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_shape, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("shape", attrValue);
-    public static THtmlTag _src<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_src, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("src", attrValue);
-    public static THtmlTag _crossorigin<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_crossorigin, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("crossorigin", attrValue);
-    public static THtmlTag _preload<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_preload, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("preload", attrValue);
-    public static THtmlTag _autoplay<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_autoplay, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("autoplay", attrValue);
-    public static THtmlTag _loop<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_loop, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("loop", attrValue);
-    public static THtmlTag _muted<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_muted, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("muted", attrValue);
-    public static THtmlTag _controls<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_controls, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("controls", attrValue);
-    public static THtmlTag _cite<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_cite, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("cite", attrValue);
-    public static THtmlTag _onafterprint<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onafterprint, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onafterprint", attrValue);
-    public static THtmlTag _onbeforeprint<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onbeforeprint, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onbeforeprint", attrValue);
-    public static THtmlTag _onbeforeunload<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onbeforeunload, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onbeforeunload", attrValue);
-    public static THtmlTag _onhashchange<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onhashchange, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onhashchange", attrValue);
-    public static THtmlTag _onlanguagechange<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onlanguagechange, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onlanguagechange", attrValue);
-    public static THtmlTag _onmessage<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onmessage, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onmessage", attrValue);
-    public static THtmlTag _onmessageerror<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onmessageerror, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onmessageerror", attrValue);
-    public static THtmlTag _onoffline<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onoffline, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onoffline", attrValue);
-    public static THtmlTag _ononline<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_ononline, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("ononline", attrValue);
-    public static THtmlTag _onpagehide<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onpagehide, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onpagehide", attrValue);
-    public static THtmlTag _onpageshow<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onpageshow, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onpageshow", attrValue);
-    public static THtmlTag _onpopstate<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onpopstate, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onpopstate", attrValue);
-    public static THtmlTag _onrejectionhandled<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onrejectionhandled, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onrejectionhandled", attrValue);
-    public static THtmlTag _onstorage<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onstorage, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onstorage", attrValue);
-    public static THtmlTag _onunhandledrejection<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onunhandledrejection, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onunhandledrejection", attrValue);
-    public static THtmlTag _onunload<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_onunload, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("onunload", attrValue);
-    public static THtmlTag _disabled<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_disabled, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("disabled", attrValue);
-    public static THtmlTag _form<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_form, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("form", attrValue);
-    public static THtmlTag _formaction<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_formaction, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("formaction", attrValue);
-    public static THtmlTag _formenctype<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_formenctype, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("formenctype", attrValue);
-    public static THtmlTag _formmethod<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_formmethod, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("formmethod", attrValue);
-    public static THtmlTag _formnovalidate<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_formnovalidate, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("formnovalidate", attrValue);
-    public static THtmlTag _formtarget<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_formtarget, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("formtarget", attrValue);
-    public static THtmlTag _name<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_name, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("name", attrValue);
-    public static THtmlTag _popovertarget<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_popovertarget, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("popovertarget", attrValue);
-    public static THtmlTag _popovertargetaction<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_popovertargetaction, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("popovertargetaction", attrValue);
-    public static THtmlTag _value<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_value, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("value", attrValue);
-    public static THtmlTag _width<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_width, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("width", attrValue);
-    public static THtmlTag _height<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_height, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("height", attrValue);
-    public static THtmlTag _span<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_span, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("span", attrValue);
-    public static THtmlTag _datetime<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_datetime, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("datetime", attrValue);
-    public static THtmlTag _open<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_open, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("open", attrValue);
-    public static THtmlTag _accept_charset<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_accept_charset, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("accept-charset", attrValue);
-    public static THtmlTag _action<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_action, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("action", attrValue);
-    public static THtmlTag _autocomplete<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_autocomplete, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("autocomplete", attrValue);
-    public static THtmlTag _enctype<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_enctype, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("enctype", attrValue);
-    public static THtmlTag _method<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_method, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("method", attrValue);
-    public static THtmlTag _novalidate<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_novalidate, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("novalidate", attrValue);
-    public static THtmlTag _manifest<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_manifest, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("manifest", attrValue);
-    public static THtmlTag _srcdoc<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_srcdoc, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("srcdoc", attrValue);
-    public static THtmlTag _sandbox<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_sandbox, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("sandbox", attrValue);
-    public static THtmlTag _allow<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_allow, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("allow", attrValue);
-    public static THtmlTag _allowfullscreen<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_allowfullscreen, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("allowfullscreen", attrValue);
-    public static THtmlTag _loading<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_loading, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("loading", attrValue);
-    public static THtmlTag _srcset<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_srcset, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("srcset", attrValue);
-    public static THtmlTag _sizes<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_sizes, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("sizes", attrValue);
-    public static THtmlTag _usemap<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_usemap, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("usemap", attrValue);
-    public static THtmlTag _ismap<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_ismap, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("ismap", attrValue);
-    public static THtmlTag _decoding<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_decoding, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("decoding", attrValue);
-    public static THtmlTag _fetchpriority<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_fetchpriority, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("fetchpriority", attrValue);
-    public static THtmlTag _accept<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_accept, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("accept", attrValue);
-    public static THtmlTag _checked<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_checked, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("checked", attrValue);
-    public static THtmlTag _dirname<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_dirname, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("dirname", attrValue);
-    public static THtmlTag _list<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_list, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("list", attrValue);
-    public static THtmlTag _max<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_max, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("max", attrValue);
-    public static THtmlTag _maxlength<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_maxlength, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("maxlength", attrValue);
-    public static THtmlTag _min<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_min, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("min", attrValue);
-    public static THtmlTag _minlength<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_minlength, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("minlength", attrValue);
-    public static THtmlTag _multiple<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_multiple, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("multiple", attrValue);
-    public static THtmlTag _pattern<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_pattern, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("pattern", attrValue);
-    public static THtmlTag _placeholder<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_placeholder, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("placeholder", attrValue);
-    public static THtmlTag _readonly<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_readonly, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("readonly", attrValue);
-    public static THtmlTag _required<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_required, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("required", attrValue);
-    public static THtmlTag _size<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_size, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("size", attrValue);
-    public static THtmlTag _step<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_step, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("step", attrValue);
-    public static THtmlTag _for<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_for, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("for", attrValue);
-    public static THtmlTag _as<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_as, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("as", attrValue);
-    public static THtmlTag _media<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_media, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("media", attrValue);
-    public static THtmlTag _imagesrcset<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_imagesrcset, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("imagesrcset", attrValue);
-    public static THtmlTag _imagesizes<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_imagesizes, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("imagesizes", attrValue);
-    public static THtmlTag _integrity<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_integrity, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("integrity", attrValue);
-    public static THtmlTag _blocking<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_blocking, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("blocking", attrValue);
-    public static THtmlTag _color<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_color, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("color", attrValue);
-    public static THtmlTag _http_equiv<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_http_equiv, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("http-equiv", attrValue);
-    public static THtmlTag _content<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_content, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("content", attrValue);
-    public static THtmlTag _charset<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_charset, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("charset", attrValue);
-    public static THtmlTag _low<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_low, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("low", attrValue);
-    public static THtmlTag _high<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_high, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("high", attrValue);
-    public static THtmlTag _optimum<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_optimum, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("optimum", attrValue);
-    public static THtmlTag _data<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_data, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("data", attrValue);
-    public static THtmlTag _reversed<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_reversed, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("reversed", attrValue);
-    public static THtmlTag _start<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_start, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("start", attrValue);
-    public static THtmlTag _label<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_label, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("label", attrValue);
-    public static THtmlTag _selected<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_selected, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("selected", attrValue);
-    public static THtmlTag _nomodule<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_nomodule, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("nomodule", attrValue);
-    public static THtmlTag _async<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_async, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("async", attrValue);
-    public static THtmlTag _defer<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_defer, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("defer", attrValue);
-    public static THtmlTag _colspan<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_colspan, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("colspan", attrValue);
-    public static THtmlTag _rowspan<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_rowspan, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("rowspan", attrValue);
-    public static THtmlTag _headers<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_headers, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("headers", attrValue);
-    public static THtmlTag _cols<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_cols, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("cols", attrValue);
-    public static THtmlTag _rows<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_rows, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("rows", attrValue);
-    public static THtmlTag _wrap<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_wrap, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("wrap", attrValue);
-    public static THtmlTag _scope<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_scope, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("scope", attrValue);
-    public static THtmlTag _abbr<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_abbr, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("abbr", attrValue);
-    public static THtmlTag _default<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_default, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("default", attrValue);
-    public static THtmlTag _kind<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_kind, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("kind", attrValue);
-    public static THtmlTag _srclang<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_srclang, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("srclang", attrValue);
-    public static THtmlTag _poster<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_poster, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("poster", attrValue);
-    public static THtmlTag _playsinline<THtmlTag>(this THtmlTag htmlTagExpr, string? attrValue)
-        where THtmlTag : struct, IHasAttr_playsinline, IHtmlElement<THtmlTag>
-        => htmlTagExpr.Attribute("playsinline", attrValue);
-}
