+++ src/ProgressOnderwijsUtils/Html/HtmlSpec.HtmlTags.Generated.cs
@@ -1,0 +1,344 @@
+#nullable enable
+using static ProgressOnderwijsUtils.Html.AttributeNameInterfaces;
+
+namespace ProgressOnderwijsUtils.Html;
+
+public static class Tags
+{
+
+    ///<summary>Hyperlink. See: <a href="https://html.spec.whatwg.org/#the-a-element">https://html.spec.whatwg.org/#the-a-element</a><br /></summary>
+    public static readonly HtmlTagKinds.A _a = new HtmlTagKinds.A();
+
+    ///<summary>Abbreviation. See: <a href="https://html.spec.whatwg.org/#the-abbr-element">https://html.spec.whatwg.org/#the-abbr-element</a><br /></summary>
+    public static readonly HtmlTagKinds.ABBR _abbr = new HtmlTagKinds.ABBR();
+
+    ///<summary>Contact information for a page or article element. See: <a href="https://html.spec.whatwg.org/#the-address-element">https://html.spec.whatwg.org/#the-address-element</a><br /></summary>
+    public static readonly HtmlTagKinds.ADDRESS _address = new HtmlTagKinds.ADDRESS();
+
+    ///<summary>Hyperlink or dead area on an image map. See: <a href="https://html.spec.whatwg.org/#the-area-element">https://html.spec.whatwg.org/#the-area-element</a><br /></summary>
+    public static readonly HtmlTagKinds.AREA _area = new HtmlTagKinds.AREA();
+
+    ///<summary>Self-contained syndicatable or reusable composition. See: <a href="https://html.spec.whatwg.org/#the-article-element">https://html.spec.whatwg.org/#the-article-element</a><br /></summary>
+    public static readonly HtmlTagKinds.ARTICLE _article = new HtmlTagKinds.ARTICLE();
+
+    ///<summary>Sidebar for tangentially related content. See: <a href="https://html.spec.whatwg.org/#the-aside-element">https://html.spec.whatwg.org/#the-aside-element</a><br /></summary>
+    public static readonly HtmlTagKinds.ASIDE _aside = new HtmlTagKinds.ASIDE();
+
+    ///<summary>Audio player. See: <a href="https://html.spec.whatwg.org/#the-audio-element">https://html.spec.whatwg.org/#the-audio-element</a><br /></summary>
+    public static readonly HtmlTagKinds.AUDIO _audio = new HtmlTagKinds.AUDIO();
+
+    ///<summary>Keywords. See: <a href="https://html.spec.whatwg.org/#the-b-element">https://html.spec.whatwg.org/#the-b-element</a><br /></summary>
+    public static readonly HtmlTagKinds.B _b = new HtmlTagKinds.B();
+
+    ///<summary>Base URL and default target navigable for hyperlinks and forms. See: <a href="https://html.spec.whatwg.org/#the-base-element">https://html.spec.whatwg.org/#the-base-element</a><br /></summary>
+    public static readonly HtmlTagKinds.BASE _base = new HtmlTagKinds.BASE();
+
+    ///<summary>Text directionality isolation. See: <a href="https://html.spec.whatwg.org/#the-bdi-element">https://html.spec.whatwg.org/#the-bdi-element</a><br /></summary>
+    public static readonly HtmlTagKinds.BDI _bdi = new HtmlTagKinds.BDI();
+
+    ///<summary>Text directionality formatting. See: <a href="https://html.spec.whatwg.org/#the-bdo-element">https://html.spec.whatwg.org/#the-bdo-element</a><br /></summary>
+    public static readonly HtmlTagKinds.BDO _bdo = new HtmlTagKinds.BDO();
+
+    ///<summary>A section quoted from another source. See: <a href="https://html.spec.whatwg.org/#the-blockquote-element">https://html.spec.whatwg.org/#the-blockquote-element</a><br /></summary>
+    public static readonly HtmlTagKinds.BLOCKQUOTE _blockquote = new HtmlTagKinds.BLOCKQUOTE();
+
+    ///<summary>Document body. See: <a href="https://html.spec.whatwg.org/#the-body-element">https://html.spec.whatwg.org/#the-body-element</a><br /></summary>
+    public static readonly HtmlTagKinds.BODY _body = new HtmlTagKinds.BODY();
+
+    ///<summary>Line break, e.g. in poem or postal address. See: <a href="https://html.spec.whatwg.org/#the-br-element">https://html.spec.whatwg.org/#the-br-element</a><br /></summary>
+    public static readonly HtmlTagKinds.BR _br = new HtmlTagKinds.BR();
+
+    ///<summary>Button control. See: <a href="https://html.spec.whatwg.org/#the-button-element">https://html.spec.whatwg.org/#the-button-element</a><br /></summary>
+    public static readonly HtmlTagKinds.BUTTON _button = new HtmlTagKinds.BUTTON();
+
+    ///<summary>Scriptable bitmap canvas. See: <a href="https://html.spec.whatwg.org/#the-canvas-element">https://html.spec.whatwg.org/#the-canvas-element</a><br /></summary>
+    public static readonly HtmlTagKinds.CANVAS _canvas = new HtmlTagKinds.CANVAS();
+
+    ///<summary>Table caption. See: <a href="https://html.spec.whatwg.org/#the-caption-element">https://html.spec.whatwg.org/#the-caption-element</a><br /></summary>
+    public static readonly HtmlTagKinds.CAPTION _caption = new HtmlTagKinds.CAPTION();
+
+    ///<summary>Title of a work. See: <a href="https://html.spec.whatwg.org/#the-cite-element">https://html.spec.whatwg.org/#the-cite-element</a><br /></summary>
+    public static readonly HtmlTagKinds.CITE _cite = new HtmlTagKinds.CITE();
+
+    ///<summary>Computer code. See: <a href="https://html.spec.whatwg.org/#the-code-element">https://html.spec.whatwg.org/#the-code-element</a><br /></summary>
+    public static readonly HtmlTagKinds.CODE _code = new HtmlTagKinds.CODE();
+
+    ///<summary>Table column. See: <a href="https://html.spec.whatwg.org/#the-col-element">https://html.spec.whatwg.org/#the-col-element</a><br /></summary>
+    public static readonly HtmlTagKinds.COL _col = new HtmlTagKinds.COL();
+
+    ///<summary>Group of columns in a table. See: <a href="https://html.spec.whatwg.org/#the-colgroup-element">https://html.spec.whatwg.org/#the-colgroup-element</a><br /></summary>
+    public static readonly HtmlTagKinds.COLGROUP _colgroup = new HtmlTagKinds.COLGROUP();
+
+    ///<summary>Machine-readable equivalent. See: <a href="https://html.spec.whatwg.org/#the-data-element">https://html.spec.whatwg.org/#the-data-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DATA _data = new HtmlTagKinds.DATA();
+
+    ///<summary>Container for options for combo box control. See: <a href="https://html.spec.whatwg.org/#the-datalist-element">https://html.spec.whatwg.org/#the-datalist-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DATALIST _datalist = new HtmlTagKinds.DATALIST();
+
+    ///<summary>Content for corresponding dt element(s). See: <a href="https://html.spec.whatwg.org/#the-dd-element">https://html.spec.whatwg.org/#the-dd-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DD _dd = new HtmlTagKinds.DD();
+
+    ///<summary>A removal from the document. See: <a href="https://html.spec.whatwg.org/#the-del-element">https://html.spec.whatwg.org/#the-del-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DEL _del = new HtmlTagKinds.DEL();
+
+    ///<summary>Disclosure control for hiding details. See: <a href="https://html.spec.whatwg.org/#the-details-element">https://html.spec.whatwg.org/#the-details-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DETAILS _details = new HtmlTagKinds.DETAILS();
+
+    ///<summary>Defining instance. See: <a href="https://html.spec.whatwg.org/#the-dfn-element">https://html.spec.whatwg.org/#the-dfn-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DFN _dfn = new HtmlTagKinds.DFN();
+
+    ///<summary>Dialog box or window. See: <a href="https://html.spec.whatwg.org/#the-dialog-element">https://html.spec.whatwg.org/#the-dialog-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DIALOG _dialog = new HtmlTagKinds.DIALOG();
+
+    ///<summary>Generic flow container, or container for name-value groups in dl elements. See: <a href="https://html.spec.whatwg.org/#the-div-element">https://html.spec.whatwg.org/#the-div-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DIV _div = new HtmlTagKinds.DIV();
+
+    ///<summary>Association list consisting of zero or more name-value groups. See: <a href="https://html.spec.whatwg.org/#the-dl-element">https://html.spec.whatwg.org/#the-dl-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DL _dl = new HtmlTagKinds.DL();
+
+    ///<summary>Legend for corresponding dd element(s). See: <a href="https://html.spec.whatwg.org/#the-dt-element">https://html.spec.whatwg.org/#the-dt-element</a><br /></summary>
+    public static readonly HtmlTagKinds.DT _dt = new HtmlTagKinds.DT();
+
+    ///<summary>Stress emphasis. See: <a href="https://html.spec.whatwg.org/#the-em-element">https://html.spec.whatwg.org/#the-em-element</a><br /></summary>
+    public static readonly HtmlTagKinds.EM _em = new HtmlTagKinds.EM();
+
+    ///<summary>Plugin. See: <a href="https://html.spec.whatwg.org/#the-embed-element">https://html.spec.whatwg.org/#the-embed-element</a><br /></summary>
+    public static readonly HtmlTagKinds.EMBED _embed = new HtmlTagKinds.EMBED();
+
+    ///<summary>Group of form controls. See: <a href="https://html.spec.whatwg.org/#the-fieldset-element">https://html.spec.whatwg.org/#the-fieldset-element</a><br /></summary>
+    public static readonly HtmlTagKinds.FIELDSET _fieldset = new HtmlTagKinds.FIELDSET();
+
+    ///<summary>Caption for figure. See: <a href="https://html.spec.whatwg.org/#the-figcaption-element">https://html.spec.whatwg.org/#the-figcaption-element</a><br /></summary>
+    public static readonly HtmlTagKinds.FIGCAPTION _figcaption = new HtmlTagKinds.FIGCAPTION();
+
+    ///<summary>Figure with optional caption. See: <a href="https://html.spec.whatwg.org/#the-figure-element">https://html.spec.whatwg.org/#the-figure-element</a><br /></summary>
+    public static readonly HtmlTagKinds.FIGURE _figure = new HtmlTagKinds.FIGURE();
+
+    ///<summary>Footer for a page or section. See: <a href="https://html.spec.whatwg.org/#the-footer-element">https://html.spec.whatwg.org/#the-footer-element</a><br /></summary>
+    public static readonly HtmlTagKinds.FOOTER _footer = new HtmlTagKinds.FOOTER();
+
+    ///<summary>User-submittable form. See: <a href="https://html.spec.whatwg.org/#the-form-element">https://html.spec.whatwg.org/#the-form-element</a><br /></summary>
+    public static readonly HtmlTagKinds.FORM _form = new HtmlTagKinds.FORM();
+
+    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.H1 _h1 = new HtmlTagKinds.H1();
+
+    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.H2 _h2 = new HtmlTagKinds.H2();
+
+    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.H3 _h3 = new HtmlTagKinds.H3();
+
+    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.H4 _h4 = new HtmlTagKinds.H4();
+
+    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.H5 _h5 = new HtmlTagKinds.H5();
+
+    ///<summary>Heading. See: <a href="https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements">https://html.spec.whatwg.org/#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.H6 _h6 = new HtmlTagKinds.H6();
+
+    ///<summary>Container for document metadata. See: <a href="https://html.spec.whatwg.org/#the-head-element">https://html.spec.whatwg.org/#the-head-element</a><br /></summary>
+    public static readonly HtmlTagKinds.HEAD _head = new HtmlTagKinds.HEAD();
+
+    ///<summary>Introductory or navigational aids for a page or section. See: <a href="https://html.spec.whatwg.org/#the-header-element">https://html.spec.whatwg.org/#the-header-element</a><br /></summary>
+    public static readonly HtmlTagKinds.HEADER _header = new HtmlTagKinds.HEADER();
+
+    ///<summary>Heading container. See: <a href="https://html.spec.whatwg.org/#the-hgroup-element">https://html.spec.whatwg.org/#the-hgroup-element</a><br /></summary>
+    public static readonly HtmlTagKinds.HGROUP _hgroup = new HtmlTagKinds.HGROUP();
+
+    ///<summary>Thematic break. See: <a href="https://html.spec.whatwg.org/#the-hr-element">https://html.spec.whatwg.org/#the-hr-element</a><br /></summary>
+    public static readonly HtmlTagKinds.HR _hr = new HtmlTagKinds.HR();
+
+    ///<summary>Root element. See: <a href="https://html.spec.whatwg.org/#the-html-element">https://html.spec.whatwg.org/#the-html-element</a><br /></summary>
+    public static readonly HtmlTagKinds.HTML _html = new HtmlTagKinds.HTML();
+
+    ///<summary>Alternate voice. See: <a href="https://html.spec.whatwg.org/#the-i-element">https://html.spec.whatwg.org/#the-i-element</a><br /></summary>
+    public static readonly HtmlTagKinds.I _i = new HtmlTagKinds.I();
+
+    ///<summary>Child navigable. See: <a href="https://html.spec.whatwg.org/#the-iframe-element">https://html.spec.whatwg.org/#the-iframe-element</a><br /></summary>
+    public static readonly HtmlTagKinds.IFRAME _iframe = new HtmlTagKinds.IFRAME();
+
+    ///<summary>Image. See: <a href="https://html.spec.whatwg.org/#the-img-element">https://html.spec.whatwg.org/#the-img-element</a><br /></summary>
+    public static readonly HtmlTagKinds.IMG _img = new HtmlTagKinds.IMG();
+
+    ///<summary>Form control. See: <a href="https://html.spec.whatwg.org/#the-input-element">https://html.spec.whatwg.org/#the-input-element</a><br /></summary>
+    public static readonly HtmlTagKinds.INPUT _input = new HtmlTagKinds.INPUT();
+
+    ///<summary>An addition to the document. See: <a href="https://html.spec.whatwg.org/#the-ins-element">https://html.spec.whatwg.org/#the-ins-element</a><br /></summary>
+    public static readonly HtmlTagKinds.INS _ins = new HtmlTagKinds.INS();
+
+    ///<summary>User input. See: <a href="https://html.spec.whatwg.org/#the-kbd-element">https://html.spec.whatwg.org/#the-kbd-element</a><br /></summary>
+    public static readonly HtmlTagKinds.KBD _kbd = new HtmlTagKinds.KBD();
+
+    ///<summary>Caption for a form control. See: <a href="https://html.spec.whatwg.org/#the-label-element">https://html.spec.whatwg.org/#the-label-element</a><br /></summary>
+    public static readonly HtmlTagKinds.LABEL _label = new HtmlTagKinds.LABEL();
+
+    ///<summary>Caption for fieldset. See: <a href="https://html.spec.whatwg.org/#the-legend-element">https://html.spec.whatwg.org/#the-legend-element</a><br /></summary>
+    public static readonly HtmlTagKinds.LEGEND _legend = new HtmlTagKinds.LEGEND();
+
+    ///<summary>List item. See: <a href="https://html.spec.whatwg.org/#the-li-element">https://html.spec.whatwg.org/#the-li-element</a><br /></summary>
+    public static readonly HtmlTagKinds.LI _li = new HtmlTagKinds.LI();
+
+    ///<summary>Link metadata. See: <a href="https://html.spec.whatwg.org/#the-link-element">https://html.spec.whatwg.org/#the-link-element</a><br /></summary>
+    public static readonly HtmlTagKinds.LINK _link = new HtmlTagKinds.LINK();
+
+    ///<summary>Container for the dominant contents of the document. See: <a href="https://html.spec.whatwg.org/#the-main-element">https://html.spec.whatwg.org/#the-main-element</a><br /></summary>
+    public static readonly HtmlTagKinds.MAIN _main = new HtmlTagKinds.MAIN();
+
+    ///<summary>Image map. See: <a href="https://html.spec.whatwg.org/#the-map-element">https://html.spec.whatwg.org/#the-map-element</a><br /></summary>
+    public static readonly HtmlTagKinds.MAP _map = new HtmlTagKinds.MAP();
+
+    ///<summary>Highlight. See: <a href="https://html.spec.whatwg.org/#the-mark-element">https://html.spec.whatwg.org/#the-mark-element</a><br /></summary>
+    public static readonly HtmlTagKinds.MARK _mark = new HtmlTagKinds.MARK();
+
+    ///<summary>Menu of commands. See: <a href="https://html.spec.whatwg.org/#the-menu-element">https://html.spec.whatwg.org/#the-menu-element</a><br /></summary>
+    public static readonly HtmlTagKinds.MENU _menu = new HtmlTagKinds.MENU();
+
+    ///<summary>Text metadata. See: <a href="https://html.spec.whatwg.org/#the-meta-element">https://html.spec.whatwg.org/#the-meta-element</a><br /></summary>
+    public static readonly HtmlTagKinds.META _meta = new HtmlTagKinds.META();
+
+    ///<summary>Gauge. See: <a href="https://html.spec.whatwg.org/#the-meter-element">https://html.spec.whatwg.org/#the-meter-element</a><br /></summary>
+    public static readonly HtmlTagKinds.METER _meter = new HtmlTagKinds.METER();
+
+    ///<summary>Section with navigational links. See: <a href="https://html.spec.whatwg.org/#the-nav-element">https://html.spec.whatwg.org/#the-nav-element</a><br /></summary>
+    public static readonly HtmlTagKinds.NAV _nav = new HtmlTagKinds.NAV();
+
+    ///<summary>Fallback content for script. See: <a href="https://html.spec.whatwg.org/#the-noscript-element">https://html.spec.whatwg.org/#the-noscript-element</a><br /></summary>
+    public static readonly HtmlTagKinds.NOSCRIPT _noscript = new HtmlTagKinds.NOSCRIPT();
+
+    ///<summary>Image, child navigable, or plugin. See: <a href="https://html.spec.whatwg.org/#the-object-element">https://html.spec.whatwg.org/#the-object-element</a><br /></summary>
+    public static readonly HtmlTagKinds.OBJECT _object = new HtmlTagKinds.OBJECT();
+
+    ///<summary>Ordered list. See: <a href="https://html.spec.whatwg.org/#the-ol-element">https://html.spec.whatwg.org/#the-ol-element</a><br /></summary>
+    public static readonly HtmlTagKinds.OL _ol = new HtmlTagKinds.OL();
+
+    ///<summary>Group of options in a list box. See: <a href="https://html.spec.whatwg.org/#the-optgroup-element">https://html.spec.whatwg.org/#the-optgroup-element</a><br /></summary>
+    public static readonly HtmlTagKinds.OPTGROUP _optgroup = new HtmlTagKinds.OPTGROUP();
+
+    ///<summary>Option in a list box or combo box control. See: <a href="https://html.spec.whatwg.org/#the-option-element">https://html.spec.whatwg.org/#the-option-element</a><br /></summary>
+    public static readonly HtmlTagKinds.OPTION _option = new HtmlTagKinds.OPTION();
+
+    ///<summary>Calculated output value. See: <a href="https://html.spec.whatwg.org/#the-output-element">https://html.spec.whatwg.org/#the-output-element</a><br /></summary>
+    public static readonly HtmlTagKinds.OUTPUT _output = new HtmlTagKinds.OUTPUT();
+
+    ///<summary>Paragraph. See: <a href="https://html.spec.whatwg.org/#the-p-element">https://html.spec.whatwg.org/#the-p-element</a><br /></summary>
+    public static readonly HtmlTagKinds.P _p = new HtmlTagKinds.P();
+
+    ///<summary>Image. See: <a href="https://html.spec.whatwg.org/#the-picture-element">https://html.spec.whatwg.org/#the-picture-element</a><br /></summary>
+    public static readonly HtmlTagKinds.PICTURE _picture = new HtmlTagKinds.PICTURE();
+
+    ///<summary>Block of preformatted text. See: <a href="https://html.spec.whatwg.org/#the-pre-element">https://html.spec.whatwg.org/#the-pre-element</a><br /></summary>
+    public static readonly HtmlTagKinds.PRE _pre = new HtmlTagKinds.PRE();
+
+    ///<summary>Progress bar. See: <a href="https://html.spec.whatwg.org/#the-progress-element">https://html.spec.whatwg.org/#the-progress-element</a><br /></summary>
+    public static readonly HtmlTagKinds.PROGRESS _progress = new HtmlTagKinds.PROGRESS();
+
+    ///<summary>Quotation. See: <a href="https://html.spec.whatwg.org/#the-q-element">https://html.spec.whatwg.org/#the-q-element</a><br /></summary>
+    public static readonly HtmlTagKinds.Q _q = new HtmlTagKinds.Q();
+
+    ///<summary>Parenthesis for ruby annotation text. See: <a href="https://html.spec.whatwg.org/#the-rp-element">https://html.spec.whatwg.org/#the-rp-element</a><br /></summary>
+    public static readonly HtmlTagKinds.RP _rp = new HtmlTagKinds.RP();
+
+    ///<summary>Ruby annotation text. See: <a href="https://html.spec.whatwg.org/#the-rt-element">https://html.spec.whatwg.org/#the-rt-element</a><br /></summary>
+    public static readonly HtmlTagKinds.RT _rt = new HtmlTagKinds.RT();
+
+    ///<summary>Ruby annotation(s). See: <a href="https://html.spec.whatwg.org/#the-ruby-element">https://html.spec.whatwg.org/#the-ruby-element</a><br /></summary>
+    public static readonly HtmlTagKinds.RUBY _ruby = new HtmlTagKinds.RUBY();
+
+    ///<summary>Inaccurate text. See: <a href="https://html.spec.whatwg.org/#the-s-element">https://html.spec.whatwg.org/#the-s-element</a><br /></summary>
+    public static readonly HtmlTagKinds.S _s = new HtmlTagKinds.S();
+
+    ///<summary>Computer output. See: <a href="https://html.spec.whatwg.org/#the-samp-element">https://html.spec.whatwg.org/#the-samp-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SAMP _samp = new HtmlTagKinds.SAMP();
+
+    ///<summary>Embedded script. See: <a href="https://html.spec.whatwg.org/#the-script-element">https://html.spec.whatwg.org/#the-script-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SCRIPT _script = new HtmlTagKinds.SCRIPT();
+
+    ///<summary>Container for search controls. See: <a href="https://html.spec.whatwg.org/#the-search-element">https://html.spec.whatwg.org/#the-search-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SEARCH _search = new HtmlTagKinds.SEARCH();
+
+    ///<summary>Generic document or application section. See: <a href="https://html.spec.whatwg.org/#the-section-element">https://html.spec.whatwg.org/#the-section-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SECTION _section = new HtmlTagKinds.SECTION();
+
+    ///<summary>List box control. See: <a href="https://html.spec.whatwg.org/#the-select-element">https://html.spec.whatwg.org/#the-select-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SELECT _select = new HtmlTagKinds.SELECT();
+
+    ///<summary>Shadow tree slot. See: <a href="https://html.spec.whatwg.org/#the-slot-element">https://html.spec.whatwg.org/#the-slot-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SLOT _slot = new HtmlTagKinds.SLOT();
+
+    ///<summary>Side comment. See: <a href="https://html.spec.whatwg.org/#the-small-element">https://html.spec.whatwg.org/#the-small-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SMALL _small = new HtmlTagKinds.SMALL();
+
+    ///<summary>Image source for img or media source for video or audio. See: <a href="https://html.spec.whatwg.org/#the-source-element">https://html.spec.whatwg.org/#the-source-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SOURCE _source = new HtmlTagKinds.SOURCE();
+
+    ///<summary>Generic phrasing container. See: <a href="https://html.spec.whatwg.org/#the-span-element">https://html.spec.whatwg.org/#the-span-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SPAN _span = new HtmlTagKinds.SPAN();
+
+    ///<summary>Importance. See: <a href="https://html.spec.whatwg.org/#the-strong-element">https://html.spec.whatwg.org/#the-strong-element</a><br /></summary>
+    public static readonly HtmlTagKinds.STRONG _strong = new HtmlTagKinds.STRONG();
+
+    ///<summary>Embedded styling information. See: <a href="https://html.spec.whatwg.org/#the-style-element">https://html.spec.whatwg.org/#the-style-element</a><br /></summary>
+    public static readonly HtmlTagKinds.STYLE _style = new HtmlTagKinds.STYLE();
+
+    ///<summary>Subscript. See: <a href="https://html.spec.whatwg.org/#the-sub-and-sup-elements">https://html.spec.whatwg.org/#the-sub-and-sup-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.SUB _sub = new HtmlTagKinds.SUB();
+
+    ///<summary>Caption for details. See: <a href="https://html.spec.whatwg.org/#the-summary-element">https://html.spec.whatwg.org/#the-summary-element</a><br /></summary>
+    public static readonly HtmlTagKinds.SUMMARY _summary = new HtmlTagKinds.SUMMARY();
+
+    ///<summary>Superscript. See: <a href="https://html.spec.whatwg.org/#the-sub-and-sup-elements">https://html.spec.whatwg.org/#the-sub-and-sup-elements</a><br /></summary>
+    public static readonly HtmlTagKinds.SUP _sup = new HtmlTagKinds.SUP();
+
+    ///<summary>Table. See: <a href="https://html.spec.whatwg.org/#the-table-element">https://html.spec.whatwg.org/#the-table-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TABLE _table = new HtmlTagKinds.TABLE();
+
+    ///<summary>Group of rows in a table. See: <a href="https://html.spec.whatwg.org/#the-tbody-element">https://html.spec.whatwg.org/#the-tbody-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TBODY _tbody = new HtmlTagKinds.TBODY();
+
+    ///<summary>Table cell. See: <a href="https://html.spec.whatwg.org/#the-td-element">https://html.spec.whatwg.org/#the-td-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TD _td = new HtmlTagKinds.TD();
+
+    ///<summary>Template. See: <a href="https://html.spec.whatwg.org/#the-template-element">https://html.spec.whatwg.org/#the-template-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TEMPLATE _template = new HtmlTagKinds.TEMPLATE();
+
+    ///<summary>Multiline text controls. See: <a href="https://html.spec.whatwg.org/#the-textarea-element">https://html.spec.whatwg.org/#the-textarea-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TEXTAREA _textarea = new HtmlTagKinds.TEXTAREA();
+
+    ///<summary>Group of footer rows in a table. See: <a href="https://html.spec.whatwg.org/#the-tfoot-element">https://html.spec.whatwg.org/#the-tfoot-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TFOOT _tfoot = new HtmlTagKinds.TFOOT();
+
+    ///<summary>Table header cell. See: <a href="https://html.spec.whatwg.org/#the-th-element">https://html.spec.whatwg.org/#the-th-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TH _th = new HtmlTagKinds.TH();
+
+    ///<summary>Group of heading rows in a table. See: <a href="https://html.spec.whatwg.org/#the-thead-element">https://html.spec.whatwg.org/#the-thead-element</a><br /></summary>
+    public static readonly HtmlTagKinds.THEAD _thead = new HtmlTagKinds.THEAD();
+
+    ///<summary>Machine-readable equivalent of date- or time-related data. See: <a href="https://html.spec.whatwg.org/#the-time-element">https://html.spec.whatwg.org/#the-time-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TIME _time = new HtmlTagKinds.TIME();
+
+    ///<summary>Document title. See: <a href="https://html.spec.whatwg.org/#the-title-element">https://html.spec.whatwg.org/#the-title-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TITLE _title = new HtmlTagKinds.TITLE();
+
+    ///<summary>Table row. See: <a href="https://html.spec.whatwg.org/#the-tr-element">https://html.spec.whatwg.org/#the-tr-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TR _tr = new HtmlTagKinds.TR();
+
+    ///<summary>Timed text track. See: <a href="https://html.spec.whatwg.org/#the-track-element">https://html.spec.whatwg.org/#the-track-element</a><br /></summary>
+    public static readonly HtmlTagKinds.TRACK _track = new HtmlTagKinds.TRACK();
+
+    ///<summary>Unarticulated annotation. See: <a href="https://html.spec.whatwg.org/#the-u-element">https://html.spec.whatwg.org/#the-u-element</a><br /></summary>
+    public static readonly HtmlTagKinds.U _u = new HtmlTagKinds.U();
+
+    ///<summary>List. See: <a href="https://html.spec.whatwg.org/#the-ul-element">https://html.spec.whatwg.org/#the-ul-element</a><br /></summary>
+    public static readonly HtmlTagKinds.UL _ul = new HtmlTagKinds.UL();
+
+    ///<summary>Variable. See: <a href="https://html.spec.whatwg.org/#the-var-element">https://html.spec.whatwg.org/#the-var-element</a><br /></summary>
+    public static readonly HtmlTagKinds.VAR _var = new HtmlTagKinds.VAR();
+
+    ///<summary>Video player. See: <a href="https://html.spec.whatwg.org/#the-video-element">https://html.spec.whatwg.org/#the-video-element</a><br /></summary>
+    public static readonly HtmlTagKinds.VIDEO _video = new HtmlTagKinds.VIDEO();
+
+    ///<summary>Line breaking opportunity. See: <a href="https://html.spec.whatwg.org/#the-wbr-element">https://html.spec.whatwg.org/#the-wbr-element</a><br /></summary>
+    public static readonly HtmlTagKinds.WBR _wbr = new HtmlTagKinds.WBR();
+}
